tasty-dev/requirements.txt
pytest>=9.0.0
pyyaml>=6.0.0
